+++ ios/Shiori/Features/Home/HomeAPI.swift
@@ -18,7 +18,6 @@ enum HomeAPI {
             suggestions: dashboard.suggestions.map { $0.fragments.dashboardBookCard.asBook(status: .toRead) },
             lastFinished: dashboard.lastFinished?.fragments.dashboardBookCard.asBook(status: .read),
             booksRead: .init(current: dashboard.booksRead.current, previous: dashboard.booksRead.previous),
-            monthlyListeningHours: .init(current: dashboard.monthlyListeningHours.current, previous: dashboard.monthlyListeningHours.previous),
             toReadCount: dashboard.toReadCount,
             readCount: dashboard.readCount,
             monthsToClearPile: dashboard.monthsToClearPile,